requirements.txt
aiohttp==3.7.4
alpaca-trade-api==1.5.1
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
backcall==0.2.0
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
chromedriver-autoinstaller==0.3.1
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
deprecation==2.1.0
entrypoints==0.3
fonttools==4.28.5
h11==0.12.0
idna==3.3
ipykernel==6.6.0
ipython==7.30.1
jedi==0.18.1
jupyter-client==7.1.0
jupyter-core==4.9.1
kiwisolver==1.3.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
mplfinance==0.12.8b6
msgpack==1.0.2
multidict==6.0.2
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.22.0
outcome==1.1.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.5.1
prompt-toolkit==3.0.24
pycparser==2.21
Pygments==2.11.1
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
pywin32==303
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
tvdatafeed==1.2.0
typing_extensions==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
websocket-client==1.2.3
websockets==10.2
wsproto==1.0.0
yarl==1.7.2
